Implemented XML Parsing for Unit Type
Would like to make XML Parser universal, with input parameters to
specify which XML file to parse, as well as which specific index of the
repeating XML section to parse. Might want to also have a parameter for
which fields within the repeating XML section (at specific index or not)
to parse.