Added Player, City data types.
Ran into an issue with XMLparser if I want to use DataTypes as data
members of other DataTypes. Could be resolved by inclusion of parameter
parsing in the respective XML files.

FOR EXAMPLE, <resource type = "gold">5</resource> would be parsed as a
Resource DataType, whose "amount" data member would be set to 5. Similar
implementation could be done for <requiredBuilding type =
"Silk">2</requiredBuidling>. Might want to change the RequiredBuiding
tag to a String, rather than an Integer. Just a thought.