Added CheckInLog, MarketLog data types.
Added InvalidCheckInLogRemainderException.
Added InvalidMarketLogResourceValueException.
These two classes have an implementation of getLastInstance(), which
should return only the last instance of CheckInLog or MarketLog,
respectively. CheckInlog's getLastInstance method has parameters for
Player ID and City ID, since there isn't a case yet where we want to
just grab the last instance of CheckInLog, regardless of Player & City.

TO DO: Turn GameWindow into a usable GUI for launching the application
(as eventually planned to be). This could be a push-button-to-test kind
of thing for now, since there isn't any actual gameplay coded yet. That
comes after CDT creation, XMLwriter, and (possibly) JDBMS
implementation. *Thanks, Justin, for the push to look into this.