Added functionality to a few CDTs. Re-worked Screen.render method.
Screen.render() now takes an additional int as first parameter which
indicates the size of the tile to be rendered.

* TODO: Add 4x4 tile rendering for SMALL font, other (future) smaller
tiles.

Added beginning of animation to Player.render() for IDLE as well as
ACTION animations.

Added <SHIFT> Key to InputHandler so we can indicate when the Player
should be performing a more ENERGETIC animation (RUNNING, ATTACKING,
DEFENDING, etc.)